# api/docs/swagger.yaml
basePath: /services
definitions:
...
  api_internal_models.User:
    properties:
      is_admin:
        type: boolean
      last_login:
        type: string
      uid:
        type: string
      username:
        type: string
    type: object
  api_internal_models.UserPermission:
    properties:
      oc_group:
        type: boolean
      oc_user:
        type: boolean
      occtl:
        type: boolean
      statistic:
        type: boolean
      system:
        type: boolean
    type: object
...
  api_internal_routes_middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
  api_pkg_utils.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
  api_pkg_utils.ResponsePagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
...
  internal_services_staff_management.CreateStaffRequest:
    properties:
      permission:
        properties:
          oc_group:
            type: boolean
          oc_user:
            type: boolean
          occtl:
            type: boolean
          statistic:
            type: boolean
          system:
            type: boolean
        type: object
      user:
        properties:
          password:
            example: doe123456
            maxLength: 16
            minLength: 2
            type: string
          username:
            example: john_doe
            maxLength: 16
            minLength: 2
            type: string
        required:
        - password
        - username
        type: object
    type: object
  internal_services_staff_management.StaffsResponse:
    properties:
      meta:
        $ref: '#/definitions/api_pkg_utils.ResponsePagination'
      staffs:
        items:
          $ref: '#/definitions/api_internal_models.User'
        type: array
    type: object
  internal_services_staff_management.UpdateStaffPasswordRequest:
    properties:
      password:
        example: doe123456
        maxLength: 16
        minLength: 2
        type: string
    required:
    - password
    type: object
...
  internal_services_user.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 16
        minLength: 2
        type: string
      old_password:
        maxLength: 16
        minLength: 2
        type: string
    required:
    - new_password
    - old_password
    type: object
  internal_services_user.CreateAdminUserRequest:
    properties:
      password:
        example: doe123456
        maxLength: 16
        minLength: 2
        type: string
      username:
        example: john_doe
        maxLength: 16
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  internal_services_user.CreateAdminUserResponse:
    properties:
      token:
        type: string
    type: object
  internal_services_user.LoginRequest:
    properties:
      password:
        maxLength: 16
        minLength: 2
        type: string
      remember_me:
        type: boolean
      username:
        maxLength: 16
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  internal_services_user.LoginResponse:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /api/v1/staffs:
    get:
      consumes:
      - application/json
      description: List of Staffs
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Field to order by
        in: query
        name: pager
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_staff_management.StaffsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
        "403":
          description: Forbidden
      summary: Staffs
      tags:
      - Staff Management
    post:
      consumes:
      - application/json
      description: Create Staff with Permission
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Staff user and permission body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_staff_management.CreateStaffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
        "403":
          description: Forbidden
      summary: Create Staff
      tags:
      - Staff Management
  /api/v1/staffs/:uid:
    delete:
      consumes:
      - application/json
      description: Delete Staff
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
        "403":
          description: Forbidden
      summary: Delete Staff
      tags:
      - Staff Management
    patch:
      consumes:
      - application/json
      description: Update Staff Password
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      - description: Staff password update body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_staff_management.UpdateStaffPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
        "403":
          description: Forbidden
      summary: Update Staff Password
      tags:
      - Staff Management
  /api/v1/staffs/:uid/permission:
    get:
      consumes:
      - application/json
      description: Staff Permission
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_internal_models.UserPermission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
        "403":
          description: Forbidden
      summary: Staff Permission
      tags:
      - Staff Management
    patch:
      consumes:
      - application/json
      description: Update Staff Permission
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      - description: Staff permission body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_internal_models.UserPermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
        "403":
          description: Forbidden
      summary: Update Staff Permission
      tags:
      - Staff Management
...
  /api/v1/user/admin:
    post:
      consumes:
      - application/json
      description: Create Superuser in initializing step
      parameters:
      - description: check secret key from file 'init_secret'
        in: query
        name: secret
        required: true
        type: string
      - description: admin user body data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_user.CreateAdminUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_user.CreateAdminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
      summary: Create Superuser
      tags:
      - Site User
  /api/v1/user/change_password:
    post:
      consumes:
      - application/json
      description: ChangePassword Admin or Staff User change password with send old
        and new password
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_user.ChangePasswordRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: ChangePassword Admin or Staff User change password
      tags:
      - Site User
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      description: Login Admin or Staff User to get Token for request
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_user.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Login Admin or Staff User
      tags:
      - Site User
  /api/v1/user/logout:
    delete:
      consumes:
      - application/json
      description: Logout Admin or Staff User
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Logout Admin or Staff User
      tags:
      - Site User
swagger: "2.0"